✨ feat(core): Add wait_onclean parameter to send method
This commit adds the wait_onclean parameter to the send method in the EventBus class. This parameter allows the event to be sent as wait_onclean, which means that the event will be processed and waited to finish. This is useful for events that should be completed on a graceful shutdown or interruption.